+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
  * Added `WhsAppPackage` task to `whsbuild.yml` schema and `Invoke-WhsCIBuild`. This task creates application deployment packages.
  * Created `Install-WhsCITool` function for installing tools (currently only PowerShell modules) needed by WhsCI functions.
  * Created `Invoke-WhsCIPester3Task` function for running tests with Pester 3.
+ * Created `Invoke-WhsCINodeTask` function for running Node.js builds.
 
 
 # 0.1.0